backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
openai>=1.0.0
psycopg2-binary
pgvector
numpy
pydantic
sqlalchemy
aiofiles
python-multipart
python-socketio
fastapi-socketio
